Adds failing test for edge case
`symfony/css-selector` version `v3.4.37` was released on 2020/01/01. But version `v5.0.2` was released on 2019/11/18. The `LatestAsOf` method is incorrectly assuming that the most recently published package is the latest one. It instead should be returning the highest version number that was published as of the specified date. That would be `v5.0.2` in this case.