Switches to a different "unsupported" repo
The `gocd` repository has a few really long file names that are bumping up against a know issue with libgit2 and libget2net. This should address the specific issue that's highlighted by #205 (Acceptance Test Fails in Windows Development Environment), but it does not address the underlying issue with libgit2 and libgit2net.