Preparing Freshli for CLI Breakout
Moved Program and OutputFormatter and related tests to CLI folder to prepare for the CLI breakout to a different repository. A README was also initially created for the CLI component. Once the breakout happens this README should be refined.
The ManifestFinder and FileHistoryFinder Registry was also moved to the Runner to separate this from the CLI. If we need further extend the FileHistoryFinder types, we can have that declared for a specific runner. I relocated this here to cut down on code duplication as these two line are repeated in multiple locations.