Created Solution Structure for Freshli-CLI
This is the initial main commit to establish the structure for the freshli-cli project. The file within Freshli.CLI and Freshli.CLI.Test are migrated from the original freshli project and contain the necessary git history.
The cs projects Freshli.CLI and Freshli.CLI.Test are the projects to be used for this repository. The Freshli project has been included in the solution and is imported using a git submodule. This submodule is currently on the cli-breakout branch of freshli and will need to be updated once that branch is merged into main.
The README has also been adjusted to account for the new project and dotnet build and test output.