Updated Build to use Freshli Class Library
Updated freshli-cli to point at the latest commit in the cli-breakout branch of freshli. This ensures that it is using freshli as a class library and contains the removal of the files that have migrated to freshli-cli.
Fixed the import and namespace for OutputFormatterTest.cs in Freshli.Cli.Test.csproj