feat: implement 'git clone' command
Implements #51. This command clones a repository in the cache
directory and (optionally) checks out a branch. The repository URL
and branch name are also stored in the cache database. If the
repository and branch are already checked out, the latest version
is pulled instead. Also supports overriding the Git binary used
with the '--git-path' optional argument.