refactor/test: resolve Aruba paths to simulated home directory
Move the function for resolving paths in Aruba in our own step
definitions into a dedicated support function.
Both relative paths and paths starting with ~ should be resolved
to the simulated and temporary home directory created by Aruba.
This seems to be how the built-in step definitions from Aruba/Cucumber
in Ruby seem to be interpreting paths anyway.