Replaces bash build scripts with ruby equivalents
A concern was raised at the portability of the bash version. The ruby versions are an attempt to improve portability. Since ruby is already required to run `cucumber` this should work as an improvement. The ruby-based build scripts just need to avoid using any dependencies that are located in ruby gems.