Adds support for running other linters
Add ability to run `rubocop`, `dotnet format`, and `codeclimate`. Also adds a helper function in the `bin/support` directory that knows how to stream the contents of stdout and stderr from child processes. The built-in ruby `system` function only redirects stdout. Also enables colorized output for the commands that provide a mechanism for doing so.