Calculate libyear from a given cycloneDx file in JSON format
Given a CycloneDX file in JSON. Use the dependency manager repositories
to figure out what the latest version is. With this latest version we
can request the release date for both versions and calculate the
libyear.

We need to do this for all packages, and yes that could mean a lot of
calls to the repositories. The idea is that their implementation will
use a cache to fetch info for a package once (like tags and releases
dates).