build(lint): Runs ReSharper inspections as part of `bin/lint.rb`
ReSharper's inspections detect common errors and code smells with C# code.

The `dotnet jb inspectcode` command does not return a non-zero exit code when
it detects warnings or errors. This required us to read the plain-text output to
determine if any issues were reported. We also had to specify an output file,
because the command does not output detected issues to the console. The command's
default format is an XML document, so we specified plain-text instead.

Co-Authored-By: M. Scott Ford <scott@corgibytes.com>
Co-Authored-By: Dona Pop <dona@corgibytes.com>
Co-Authored-By: Jason C. McDonald <jasonm@corgibytes.com>