test: make test pass now that nullable is enabled
Explicit nullability broke nested mocking of a DirectoryInfo
object in a mocked ScanCommandOptions object. To get around this,
I explicitly instantiated a ScanCommandOptions object with a real
relative path to the present working directory for the unit test being
executed, which is a temporary path anyway. This allowed the test to
pass again.