bug: Pulling when there's no branch defined (which is default) error
When git pull is used without branch it results in an error in this
particular bit. The arguments builder can not filter out empty
arguments.

This also revealed a bug in the code that if we call the freshli git
clone command twice (so it will pull) that it doesn't work. This needs
to be addressed still.