ci: Disable the similar code check in CodeClimate
We've upped the threshold for this check a few times. Each time it was
for boiler plate code in tests. The check itself is useful but only if
we could exclude the tests for this particular check which we can't
(yet?). Until that time, since it's losing value, disable the check
entirely.