build: reverse the order of formatters
The order in which we run formatters in bin/format.rb should be
the same as the order in which we run the associated linters in
bin/lint.rb. This ensures consistency in precedence, and allows
dotnet format (which was flagging the change made by jb cleanupcode)
the opportunity to reverse the unwanted change.

However, this is only a workaround. We still need to determine
the conflicting rules in `.editorconfig` that is causing this
formatting change in MainCommand.rb to thrash.