refactor: abstract out external process calls
Adds the Invoke.Command() and Invoke.Freshli() methods for
executing external processes with CliWrap. This eliminates
boilerplate from the rest of the project.

Changes all prior direct invocations of processes with CliWrap
to use Invoke.Command() instead.

Invoke.Freshli() method will be needed in upcoming work. This just
defines it, since it was written as part of a now-abandoned task
and worth retaining.

Co-Authored-By: Dona Pop <dona@corgibytes.com>