feat: migrates `agents detect` to application engine
This commit introduces an "application engine", implemented using the Hangfire library. This engine manages a queue that contains "application activities" and "applications events". An activity is capable of firing an event, and an event is capable of dispatching an activity.

The queue is serviced by a configurable number of background worker threads, permitting parallel execution and processing of activities and events.

This architecture is going to be used in place of the one that we've been working towards so far, where command line commands execute other command line commands and process their results. Instead, `IApplicationActivity` and `IApplicationEvent` objects are going to be created to process the pipeline of work that has been sketched out in #55.

This commit is the first step towards implementing something similar to the approach that is is outlined and proposed in #135. The exact implementation here differs slightly from what was proposed, but the spirit of the implementation is much the same.

This commit contains code that was copy and pasted from the Hangfire.AspNetCore library. The classes that are imported this way are responsible for setting up dependency injection and logging support.

Since Hangfire depends on Newtonsoft.Json, a custom class has been added to assist with the integration between the Json deserialization and the dependency injection framework that we're using. This allows the deserialization functionality to refer to object instances that are present in the inversion of control container that's managed by the dependency injection framework.

This giant commit represents a group effort.

Co-Authored-By: M. Scott Ford <scott@corgibytes.com>
Co-Authored-By: Edwin Kortman <7047894+edwinkortman@users.noreply.github.com>
Co-Authored-By: Kimarie Hoot <52430349+klhoot@users.noreply.github.com>
Co-Authored-By: donabp <50669675+donabp@users.noreply.github.com>