Extends Destroy Cache Event Handling
This commit is the result of pair programming with Scott.  Instead of
handling all the scenarios in one activity and one return event, this
adds:

ForceCacheDestroyActivity - to handle the action when the user was prompted
ConfirmationRequiedEvent - to handle the case when the --force flag was not passed
CacheDestoryFailedEvent - to handle the case when the destory cache cmd failed

This change better handles the activities and events asynchronously without introducing
pauses for user input to the activity or event handling.