Renames Functionality/Cache to Functionality/CacheDestroy
There seemed to be a conflict between the namespace for Functionality.Cache
because there was already a Cache.cs file in the Functionality dir.  Changing
the new folder name to CacheDestroy seems to clear this up.  Howevre I was
wondering if the Cache.cs file should instead be moved into the new Cache
directory so that as we port other Cache commands we can keep everything
grouped together.

This commit also includes some stylistic changes based on suggestions from
the Rider IDE.