Upgrade .NET SDK in devcontainer
This works around an issue that was causing warnings when building and errors when running `bin/lint.rb` and `bin/format.rb`.

This also makes our devcontainer image based on the official .NET SDK container base image instead of depending on an intermediate devcontainer base image. The intermediate base image did not have a version published that used the correct SDK version.