Removes commented line of code
This line of code reflects a different path to dispatch the
CloneGitRepositoryActivity.  The command runner path will go
away eventually and CloneGitRepositoryActivity will be
dispatched from the AnalysisStartedEvent which will provide
an analysisId.  The command runner path does not utilize
information from a saved analysis and therefore will not have
access to provide an analysisId.