Add the ServiceProvider as public property of the ApplicationEngine
This will allow us to access the ServiceProvider from an Activity without having to inject it anywhere. Other solution tested was to make it part of the Handle() which works as well but the concern then is if objects become really large when enqueued. By adding it to the application engine we are avoiding all of that.