Adds diagram-generator CLI program
This is intended for use by the development team, and will not be shipped as part of the application.

This program generates a Mermaid.js diagram that represents the relationships between the classes that descend from `IApplicationActivity` and `IApplicationEvent`.

An arrow with a solid line indicates that the source type creates an instance of the destination type. An arrow with a dotted line indicates that the source type is a descendant of the destination type.

The goal is to run this program as part of CI to accurately generate accurate documentation about how the different activity and event classes work together.