Introduces IConfiguration interface
The goal of this interface is to provide storage for configuration values that are needed throughout the application. For starters these are the path to the application cache and the path to the git executable, both of which can be specified via parameters when running the `analyze` command. Using this configuration interface means that we don't need to pass these values around as much, and instead can pull them out of the configuration when they are needed.