Adds way to configure web api base url
Permits setting the value via an environment variable. The value from the environment variable takes precedence over a value that is set directly.

I imagine that this value will be configurable via a command line parameter, but that will just assign the value from the command line. This precedence structure means that the environment variable value will be used instead of the value that's provided in this way. This is consistent with the way other commands handle environment variable values that can also be specified by a command parameter.