Move the workers optional from the AnalyzeCommand to the MainCommand
This is a preparation for adjusting the workers before starting the processing of Events and Activities. If we don't do this, it'll try to fetch an option from the ParseResult that doesn't exist for any commands that don't have this option. Instead of adding it everywhere, make it a general option.