Passes PackageLibYearId to activities and events
This uses the cached information from the database tables instead of
serializing and deserializing the PackageLibYear object in the activities
and events.

This also removes the HistoryStopPointId from HistoryStopData because the ID
is passed to the later activities and events, while the HistoryStopData is only
passed to the earlier activities and events.