Uses HistoryStopData to get correct checkout path for history stop point
Unit tests are not good at catching bad assumptions that developers make!

In removing the step to create a new AnalysisLocation (now
HistoryStopData) when adding a history stop point to the database, I
lost the logic that sets the correct path to use
histories/<repository ID>/<commit ID> and therefore everything after
this point was incorrectly using repositories/<repository ID>.

This change adds back the HistoryStopData object long enough that we
can get the correct Path location to store in the cached history stop
point entry in the database.