Adds verification for CachedHistoryStopPoint
This change updates the ComputeHistoryActivity unit test to add
verification of the values cached in the database.  The verification would
have caused the unit test to fail when we were incorrectly storing the
repositories path in the CachedHistoryStopPoints table instead of the
expected histories path for the commit ID of the history stop point.