Fixes issues with ComputeHistoryTest
It was almost impossible to make these tests fail, because they were using `Assert.NotStrictEqual`. I'm assuming that this assertion method was chosen by the original developer because the items in the collection that were being compared did not implement euality semantics, and the method name makes it sound like it is not strict about the definition of equal. However, this method is meant to be the opposite of the `Assert.StrictEqual` method. I make the `GitCommit` and `HistoryIntervalStop` classes `record` types, so that we have the equality semantics, and started using the normal `Assert.Equal` methods instead.

These changes uncovered some issues with the structure of the test data. The sample data appears to have been intended to be ordered with the most recent date first in the list. That was not true. There was also an error in the expected order. The method implementation appears to have been doing the correct thing, however.