Ensures that the aruba home directory is removed
This includes a `windows_safe_recursive_delete` method that works around some quirks that crop up when deleting some of the git clones that can appear in the Aruba home directory.

The contents of the `.git` directory are marked as read-only, and that causes problems on Windows by default.

Also, on Windows, if the full path to a file becomes longer than around 260 characters, then `File.stat` will report that the file does not exist. This can be worked around, by prefixing the full path with `\\?\`, thus putting the path in UNC format. See https://learn.microsoft.com/en-us/dotnet/standard/io/file-path-formats for more information on the UNC path format.