Changes app engine wait logic to use an task
This is an attempt to ensure that all processing for an task (including child tasks that are generated by the task when it is handled) is complete before the `Wait` method terminates.

This is implemented through a chain of `ICountdownEvent` objects that notify each other about changes in wait state.