Adjusts check to account for word-wrapping in command output
The `spectre.console`-based console output class wraps long lines instead of letting them stay long. This requires getting clever about how to check for output that might span a line in the output. Here the `[\s\R]+` expression is used because it will match any number or combination of whitespace or newline characters. This is placed between every word in the expected output to account for areas where the line might get wrapped.